insurance-agent-customer-crm/requirements.txt
pandas>=1.3.0
openpyxl>=3.0.0
python-docx>=0.8.10
python-pptx>=0.6.21
pdfplumber>=0.5.0
requests>=2.25.0
python-dotenv>=0.19.0
